refactor: update dependencies and improve revenue management
- **Dependencies:**
    - Promote `equatable` and `path_provider` to direct main dependencies in `pubspec.yaml` and update `equatable` to version 2.0.8.
- **Financial Module:**
    - Update `FinancialController` to return a boolean status from `addRevenue` for better error handling.
    - Enhance `add_revenue_view.dart` with success/failure snackbars and improved loading indicators.
- **Refactoring:**
    - Standardize file naming conventions: rename `forgotPassword.dart` to `forgot_password.dart` and `saveSession.dart` to `save_session.dart`.
    - Update affected imports in `route_generator.dart`.
    - Apply minor UI/UX code cleanup and formatting.
